+++ lib/utils.js
@@ -1,0 +1,3 @@
+export function getCurrentTime() {
+  return process.hrtime.bigint()
+}
